+++ .gitignore
@@ -1,1 +1,3 @@
-laundromat_app/__pycache__/*
+laundromat_app/__pycache__/
+*.sqlite3
+db.sqlite3
