+++ laundromat_app/models.py
@@ -6,23 +6,9 @@ from django.urls import reverse
 class Laundromat(models.Model):
     name = models.CharField(max_length = 100)
     location = models.CharField(max_length = 100)
+    hours = models.CharField(max_length = 100)
+    description = models.CharField(max_length = 500)
     
-   
-
-    def __str__(self):
-        return self.name
-    
-    def get_absolute_url(self):
-        return reverse('laundromat_detail', args=[str(self.id)])
-    
-
-#redoing hopefully it will push ( justin).
-
-class Laundromat(models.Model):
-    name = models.CharField(max_length = 100)
-    location = models.CharField(max_length = 100)
-   
-
     def __str__(self):
         return self.name
